separate "all packages" section to own component
